Update Introduction section and replace video background with image
