--- public/background-video.mp4
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:38be29e13a4b7d613a2ee4721ca1440bdc9b6d6f27c979062a3a733d234f8f5f
-size 38102453
