--- src/app/components/ScrollableBackground.css
@@ -1,21 +1,0 @@
-.bg-scroll-container {
-  overflow-x: auto;
-  white-space: nowrap;
-  scroll-snap-type: x mandatory;
-  -webkit-overflow-scrolling: touch;
-  display: flex;
-  width: 100%;
-  height: 100%;
-}
-
-.bg-scroll-wrapper {
-  display: flex;
-  height: 100%;
-}
-
-.bg-image {
-  flex: 0 0 auto;
-  width: 100vw;
-  height: 100%;
-  scroll-snap-align: start;
-}
