+++ src/app/page.tsx
@@ -28,7 +28,7 @@
 
 
 // src/app/page.tsx
-import Introduction from "./components/Intorduction";
+import Introduction from "./components/introduction";
 import Skills from "./components/Skill";
 import Experience from "./components/Experience";
 import Navbar from "./components/Navbar";
